Moved components folder into app
