+++ app/layout.tsx
@@ -1,5 +1,5 @@
-import Footer from "@/components/footer";
-import Navbar, { NavbarItem } from "@/components/navbar";
+import Footer from "@/app/components/footer";
+import Navbar, { NavbarItem } from "@/app/components/navbar";
 import type { Metadata } from "next";
 import { Inter_Tight } from "next/font/google";
 import "./globals.css";
